pyproject.toml
[project]
name = "heatpump-controller"
version = "0.1.0"
description = "HeatPump Controller Plus Power Control"
authors = [
    {name = "lhxxl85",email = "lhxlnsy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "pandas-stubs (>=2.3.2.250926,<3.0.0.0)",
    "pydantic (>=2.7.0,<3.0.0)",
    "pydantic-settings (>=2.4.0,<3.0.0)",
    "pymodbus (>=3.12.0,<4.0.0)",
    "colorlog (>=6.10.1,<7.0.0)",
    "pyserial (>=3.5,<4.0)",
    "redis (>=5.2.1,<6.0.0)",
    "fastapi (>=0.131.0,<0.132.0)",
    "uvicorn (>=0.34.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
